Implemented ReportingStructure.
Employee computes the number of reports, which is used to set the number
of reports in a ReportingStructure object. A test case has been created
for employee John Lennon in EmployeeServiceImplTest.java. I might make a
new file for this test if I think it makes more sense to seperate it
from the rest of tests.